requirements.txt
numpy>=1.19.2
drms>=0.6.2
python-dateutil>=2.8.2
pandas>=1.3.1
tqdm>=4.61.2
matplotlib>=3.4.3
astropy>=4.3.post1
scikit-image>=0.18.2
sunpy[all]>=3.0.1
setuptools>=49.6.0
wandb>=0.13
lightning==1.9.3
pytorch_lightning==1.9.3
mayavi
pfsspy